Sunday Night Tune Up

+++ 03_Komodo_Cafe/ProgramUI.cs
@@ -33,9 +33,11 @@ namespace _03_Komodo_Cafe
         public void Menu()
         {
             tool.MenuLine(1, "See All Menu Items");
-            tool.MenuLine(2, "Add Menu Item");
-            tool.MenuLine(3, "Delete Menu Item");
-            tool.MenuLine(4, "Exit");
+            tool.MenuLine(2, "See Specific Menu Item");
+            tool.MenuLine(3, "Add Menu Item");
+            tool.MenuLine(4, "Update Menu Item");
+            tool.MenuLine(5, "Delete Menu Item");
+            tool.MenuLine(6, "Exit");
 
             int menuResponse;
 
@@ -50,17 +52,23 @@ namespace _03_Komodo_Cafe
                         ViewAllMenuItems();
                         validEntry = true;
                         break;
-
                     case 2:
-                        AddMenuItem();
+                        GetMenuItemFromUser();
                         validEntry = true;
                         break;
                     case 3:
-                        //Delete Menu Item
-                        DeleteMenuItem();
+                        AddMenuItem();
                         validEntry = true;
                         break;
                     case 4:
+                        UpdateExistingMenuItem();
+                        validEntry = true;
+                        break;
+                    case 5:
+                        DeleteMenuItem();
+                        validEntry = true;
+                        break;
+                    case 6:
                         validEntry = true;
                         break;
                     default:
@@ -108,6 +116,53 @@ namespace _03_Komodo_Cafe
             Menu();
         }
 
+
+
+        public void GetMenuItemFromUser()
+        {
+            int pad = 18;
+            int cols = 8;
+            Console.Clear();
+
+            PullMenuItemFromList();
+
+                int selected = tool.GetIntResponse("");
+                KMenuItem item = _repo[selected - 1];
+
+            if (selected == 0)
+            {
+                tool.WriteColors("Please make a valid entry:", "Red");
+                PullMenuItemFromList();
+            }
+            else
+            {
+                tool.MessageBar(item.MealName + " ", "=", " ", pad * cols);
+                tool.PadString("#", pad / 2);
+                tool.PadString("Name", pad);
+                tool.PadString("Description", pad * 3);
+                tool.PadString("Ingredients", pad * 3);
+                tool.PadString("Price", pad / 2);
+                Console.WriteLine();
+                tool.RowDivider("=", (pad * cols));
+
+                tool.PadInt(item.MealNumber, pad/2);
+                tool.PadString(item.MealName, pad);
+                tool.PadString(item.MealDescription, pad*3);
+                tool.PadString(item.MealIngredients, pad*3);
+                tool.PadDecimal(item.MealPrice, pad/2);
+                Console.WriteLine();
+                tool.RowDivider("-", pad * cols);
+            }
+
+            Console.WriteLine();
+            Console.WriteLine();
+            Menu();
+        }
+
+
+
+
+
         public void SeedMenu()
         {
             KMenuItem newitem1 = new KMenuItem(1, "Spicy Chicken", "Spicy but light chicken sandwich and fruit.", "Chicken, buns, and komodo dragon chilis", 10.95m);
@@ -142,13 +197,87 @@ namespace _03_Komodo_Cafe
         }
 
 
+        private void UpdateExistingMenuItem()
+        {
+            PullMenuItemFromList();
+            int selection = tool.GetIntResponse("Please enter which item you want to update:");
+
+            Console.Clear();
+
+            KMenuItem oldItem = _repo[selection - 1];
+
+            if (oldItem == null)
+            {
+                tool.WriteColors("Item not found.  Press any key to continue...", "Red");
+                Console.ReadKey();
+            }
+
+            KMenuItem item = new KMenuItem(
+                oldItem.MealNumber, 
+                oldItem.MealName,
+                oldItem.MealDescription,
+                oldItem.MealIngredients,
+                oldItem.MealPrice
+                );
+
+            Console.WriteLine("Which property would you like to update:");
+
+            tool.MenuLine(1, "Item Name");
+            tool.MenuLine(2, "Item Descriptioin");
+            tool.MenuLine(3, "Item Ingredients");
+            tool.MenuLine(4, "Item Price");
+            
+            int propUpdate = tool.GetIntResponse("");
+
+            switch (propUpdate)
+            {
+                case 1:
+                    oldItem.MealName = tool.GetStringResponse("Enter new Type");
+                    break;
+                case 2:
+                    oldItem.MealDescription = tool.GetStringResponse("Enter new Make");
+                    break;
+                case 3:
+                    oldItem.MealIngredients = tool.GetStringResponse("Enter new Model");
+                    break;
+                case 4:
+                    oldItem.MealPrice = tool.GetIntResponse("Enter new Year");
+                    break;
+                default:
+                    break;
+            }
+            Menu();
+        }
+
+
+
+
 
         public void DeleteMenuItem()
         {
-            PullMenuItemFromList();
-            int selection = tool.GetIntResponse("Please enter which menu item you want to delete:");
+            Console.Clear();
 
-            _repo.RemoveAt(selection - 1);
+            Console.WriteLine("Please enter which menu item number you want to delete:");
+            Console.WriteLine();
+            PullMenuItemFromList();
+
+
+            bool validSelection = false;
+
+            while(validSelection == false)
+            {
+                int selection = tool.GetIntResponse("");
+                
+                if (selection == 0 || selection + 1 > _repo.Count)
+                {
+                    tool.WriteColors("Please make a valid selection.", 1);
+                }
+                else
+                {
+                    validSelection = true;
+                    _repo.RemoveAt(selection - 1);
+                }
+            }
 
             tool.WriteColors("Item deleted.", "Red");
